+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 - Allow array-like params in CQN.expr via `{ list: [...] }` syntax
 - Types for `cds.error()`
+- `GetRequest` and `PostRequest` types
 ### Changed
 ### Deprecated
 ### Removed